Fix warnings
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@41357 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/c64/cart/ramlink.c
@@ -800,7 +800,7 @@ static int ramlink_activate(void)
     if (rl_cardsize > rl_cardsize_old) {
         /* memset(rl_card, 0, (size_t)(rl_cardsize - rl_cardsize_old)); */
         ram_init_with_pattern(&rl_card[rl_cardsize_old],
-                              (size_t)(rl_cardsize - rl_cardsize_old), &ramparam);
+                              (unsigned int)(rl_cardsize - rl_cardsize_old), &ramparam);
     }
 
     rl_cardsize_old = rl_cardsize;
